Copy the start element also to the new root in move_siblings
In our template use case for odt documents the start element will be a text:p
(paragraph) element which may contain text:span childs. In that case the
text:span childs must be copied within a text:p element. If the text:span is not
contained in a text:p libreoffice will remove the element.