Fix variable unpacking on a data source value
If a variable that comes from the data source is unpacked, e.g. in a
for statement, it is assumed to be a tuple or list. Each value inside the
data source tuples will be extracted according to the uses of their respective
unpacked variables in the for loop's body.